+++ CHANGELOG.md
@@ -13,6 +13,10 @@ use patch releases for compatibility fixes instead.
 
 - Added `prefix` parameter to `Beaker.dataset.ls()`.
 
+### Fixed
+
+- The `FileInfo` objects from `Beaker.dataset.ls()` will now include the `digest`.
+
 ## [v1.13.2](https://github.com/allenai/beaker-py/releases/tag/v1.13.2) - 2023-01-03
 
 ### Changed